use github.com/stretchr/testify/assert from example/hello
This demonstrates that the current approach makes it easy to accidentally depend on an external module that a local module fails to specify. In this case, testify/assert was only specified by hello, but example/hello was able to "use it" and `bazel test ...:all` continued to pass even though example/hello/go.mod didn't specify it. The resolution was easy, simply run `go mod tidy` in example/hello... but this is a footgun.

Possible solutions to explore:

* Can we create per go module dependency isolation? Initial attempts at making each module it's own BZLmod created a host of other issues. For example, this turned each of those modules into a repository (prefixed with @) and simple bazel commands such as `bazel query ...` and `bazel test ...:all` do not discover them. It also created some other naming complexity, and visbility issues. 
* Can we run linting / auto-correction during the "setup/gazelle" step? This seems most straightforward, with the least caveats. At some point in time when bazel/bzlmod has a good answer for this CR repo layout we should be able to migrate.
* alternatives TBD